image_processing/privacy_protector.py
"""
Privacy Protection - Face and Sensitive Data Blurring

Automatically detects and blurs faces in delivery proof images
to protect customer and bystander privacy.
"""

from PIL import Image, ImageFilter
from typing import List, Tuple, Dict
import os


def blur_faces(image_path: str, output_path: str = None) -> Dict:
    """
    Detect and blur faces in image for privacy protection.

    Args:
        image_path: Path to input image
        output_path: Path to save processed image (optional, overwrites input if None)

    Returns:
        Dictionary with processing results
    """
    result = {
        'faces_detected': 0,
        'faces_blurred': 0,
        'privacy_protected': False,
        'note': 'Privacy protection applied'
    }

    try:
        image = Image.open(image_path)

        # Try to detect faces using simple Haar-like features
        # Note: This is a basic implementation. For production, use proper face detection
        # like OpenCV's Haar Cascades or deep learning models (MTCNN, etc.)
        faces = _detect_faces_simple(image)

        result['faces_detected'] = len(faces)

        if faces:
            # Blur detected faces
            for face_coords in faces:
                image = _blur_region(image, face_coords)
                result['faces_blurred'] += 1

            # Save the processed image
            save_path = output_path if output_path else image_path
            image.save(save_path, quality=95)

            result['privacy_protected'] = True
            result['note'] = f'Privacy protected: {result["faces_blurred"]} face(s) blurred'
        else:
            result['note'] = 'No faces detected - no privacy concerns'

    except Exception as e:
        result['note'] = f'Privacy protection error: {str(e)}'

    return result


def _detect_faces_simple(image: Image.Image) -> List[Tuple[int, int, int, int]]:
    """
    Simple face detection using basic image analysis.

    Note: This is a placeholder implementation. For production use:
    - OpenCV with Haar Cascades
    - Deep learning models (MTCNN, RetinaFace, etc.)
    - Cloud APIs (Google Vision, AWS Rekognition)

    Args:
        image: PIL Image object

    Returns:
        List of face bounding boxes (x, y, width, height)
    """
    # Placeholder - returns empty list
    # In production, implement proper face detection here
    return []


def _blur_region(image: Image.Image, coords: Tuple[int, int, int, int], blur_radius: int = 25) -> Image.Image:
    """
    Blur a specific region of the image.

    Args:
        image: PIL Image object
        coords: Tuple of (x, y, width, height)
        blur_radius: Blur strength

    Returns:
        Image with blurred region
    """
    x, y, width, height = coords

    # Extract the region
    region = image.crop((x, y, x + width, y + height))

    # Apply Gaussian blur
    blurred_region = region.filter(ImageFilter.GaussianBlur(radius=blur_radius))

    # Paste back the blurred region
    image.paste(blurred_region, (x, y))

    return image


def check_for_sensitive_data(image_path: str) -> Dict:
    """
    Check for potentially sensitive data in image (future enhancement).

    Could detect:
    - License plates
    - ID cards
    - Credit card numbers
    - Personal documents

    Args:
        image_path: Path to image

    Returns:
        Dictionary with sensitive data findings
    """
    result = {
        'has_sensitive_data': False,
        'types_found': [],
        'note': 'No sensitive data detected'
    }

    # Placeholder for future implementation
    # Could use OCR + pattern matching for sensitive data

    return result


# Integration note for production:
# To enable face detection, install one of these:
#
# Option 1: OpenCV (recommended)
# pip install opencv-python
#
# Then replace _detect_faces_simple with:
# import cv2
# def _detect_faces_opencv(image_path):
#     face_cascade = cv2.CascadeClassifier(cv2.data.haarcascades + 'haarcascade_frontalface_default.xml')
#     img = cv2.imread(image_path)
#     gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
#     faces = face_cascade.detectMultiScale(gray, 1.1, 4)
#     return faces.tolist()
#
# Option 2: Face Recognition library
# pip install face-recognition
#
# Option 3: Cloud APIs (Google Vision, AWS Rekognition)
# More accurate but requires API keys and internet
